+++ CHANGELOG.md
@@ -15,12 +15,17 @@ For each changelong entry sections are:
 
 ## [Unreleased]
 
-## [0.0.1-rc1] - 2020/07/20
+### Added
+1. Decrypt script to get the private key from a keystore account
 
+### Fixed
+1. Passphrase parametes was in plain text. Update with sensitive=true.
+2. Fix issue with passphrase when importing a key
+
+## [0.0.1-rc1] - 2020/07/20
 First release candidate of terraform-provider-ethereum
 
 ### Added
-
 1. Create ethereum accounts
 2. Update ethereum accounts
 3. Import ethereum accounts